requirements.txt
ultralytics==8.3.24
opencv-python
torch
torchvision
numpy
pillow
python-dotenv
pinecone>=5.0.0
scikit-learn
